\name{thema_black}
\alias{thema_bwblack}
\alias{theme_miniblack}
\alias{theme_classicblack}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Black themes
}
\description{
The default theme for ggplot2 is gray letters, but these are black.　Use theme() if you just need to tweak the display of an existing theme.
}
\usage{
theme_bwblack(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)

theme_miniblack(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)

theme_classicblack(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)

}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{base_size}{base font size, given in pts.}
  \item{base_family}{base font family}
  \item{base_line_size}{base size for line elements}
  \item{base_rect_size}{base size for rect elements}
}
